Adicao da Classe de Conexa e tela de Listagem de produtos

+++ LojaRoupas/LojaRoupas/Classes/Documento.cs
@@ -10,20 +10,20 @@ namespace LojaRoupas.Classes
     {
         private int id;
         private String data;
-        private float vlrTotal;
-        private float desconto;
+        private Double vlrTotal;
+        private Double desconto;
         private int qtdItens;
 
         public void setID(int id) { this.id = id; }
         public void setData(String data) { this.data = data; }
-        public void setVlrTotal(float vlrTotal) { this.vlrTotal = vlrTotal; }
-        public void setDesconto(float desconto) { this.desconto = desconto; }
+        public void setVlrTotal(Double vlrTotal) { this.vlrTotal = vlrTotal; }
+        public void setDesconto(Double desconto) { this.desconto = desconto; }
         public void setQtdItens(int qtdItens) { this.qtdItens = qtdItens; }
 
         public int getId() { return this.id; }
         public String getData() { return this.data; }
-        public float getVlrTotal() { return this.vlrTotal; }
-        public float getDesconto() { return this.desconto; }
+        public Double getVlrTotal() { return this.vlrTotal; }
+        public Double getDesconto() { return this.desconto; }
         public int getQtdItens() { return this.qtdItens; }
     }
 }
