+++ LojaRoupas/LojaRoupas/Classes/ItemCompra.cs
@@ -8,8 +8,8 @@ namespace LojaRoupas.Classes
 {
     class ItemCompra : ItemDocumento
     {
-        private float prcCusto;
-        public float getPrcCusto(){ return this.prcCusto; }
-        public void setPrcCusto(float prcCusto){ this.prcCusto = prcCusto; }
+        private Double prcCusto;
+        public Double getPrcCusto(){ return this.prcCusto; }
+        public void setPrcCusto(Double prcCusto){ this.prcCusto = prcCusto; }
     }
 }
